requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.6
django-cloudinary-storage==0.3.0
idna==3.10
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
